backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
websockets
requests
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
openai>=1.0.0
